requirements.txt
anyio==4.11.0
click==8.3.0
fastapi==0.120.2
uvicorn==0.38.0
httptools==0.7.1
websockets==15.0.1
h11==0.16.0
watchfiles==1.1.1
python-multipart==0.0.20
python-dotenv==1.2.1
PyYAML==6.0.3
idna==3.11
sniffio==1.3.1
typing_extensions==4.15.0

SQLAlchemy==2.0.36
pydantic==2.12.3
pydantic-settings==2.11.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# SOAP
fastapi-soap==0.2.0
pydantic-xml>=2.17.2
lxml>=5.3.0
zeep>=4.2.1
email-validator>=2.0.0
psycopg2-binary>=2.9.0

# Observabilidad
prometheus-fastapi-instrumentator>=7.1.0
prometheus-client>=0.21.0
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp>=1.27.0
opentelemetry-instrumentation-fastapi>=0.51b0
opentelemetry-instrumentation-asgi>=0.51b0
opentelemetry-propagator-b3>=1.27.0
opentelemetry-instrumentation-requests>=0.51b0
structlog>=24.1.0
